+++ test/runner.dart
@@ -2,7 +2,8 @@ import 'dart:async';
 import 'package:unittest/unittest.dart' show unittestConfiguration;
 
 import 'helpers/testdata.dart';
-import 'specs/elastic_search.dart' as elastic_search;
+import 'specs/elasticsearch.dart' as elasticsearch;
+import 'specs/warehouse/warehouse.dart' as warehouse;
 
 main() async {
   await setUpTestData();
@@ -11,5 +12,6 @@ main() async {
 
   unittestConfiguration.timeout = const Duration(seconds: 5);
 
-  elastic_search.main();
+  elasticsearch.main();
+  warehouse.main();
 }