+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0
+- style: Rename ElasticSearch to Elasticsearch
+- feat: Add initial Warehouse companion adapter
+
 ## 0.1.0+1
 - doc: Add backticks around examples